pyproject.toml
[project]
name = "trendfollowing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "matlabengine>=25.2.2",
    "numpy<2.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pyarrow>=23.0.0",
    "pydantic-settings>=2.12.0",
    "pykrx==1.2.3",
    "python-dotenv>=1.2.1",
    "setuptools<82",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
research = [
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.4",
    "matplotlib>=3.10.8",
]
